# samrs-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: SAMRS API
  version: "1.0.0"
  description: Sistem Manajemen Aset Rumah Sakit (MVP)
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/maintenance-schedules/{id}/documents:
    post:
      summary: Upload maintenance document
      tags: [MaintenanceDocument]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MaintenanceDocumentUpload"
      responses:
        "200":
          description: Dokumen berhasil diunggah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
    get:
      summary: List maintenance documents by schedule
      tags: [MaintenanceDocument]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Dokumen ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
...
  /api/v1/maintenance-documents/{id}:
    delete:
      summary: Delete maintenance document
      tags: [MaintenanceDocument]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Dokumen berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/documents:
    get:
      summary: List documents
      tags: [Document]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/per_page"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_dir"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/date_from"
        - $ref: "#/components/parameters/date_to"
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [sop, manual, certificate, other]
      responses:
        "200":
          description: Dokumen ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseWithMeta"
    post:
      summary: Create document
      tags: [Document]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "200":
          description: Dokumen berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/documents/{id}:
    get:
      summary: Get document by ID
      tags: [Document]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Dokumen ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
    patch:
      summary: Update document
      tags: [Document]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Dokumen berhasil diubah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
    delete:
      summary: Delete document
      tags: [Document]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Dokumen berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    per_page:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
    sort_by:
      name: sort_by
      in: query
      schema:
        type: string
    sort_dir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
    search:
      name: search
      in: query
      schema:
        type: string
    date_from:
      name: date_from
      in: query
      schema:
        type: string
        format: date
    date_to:
      name: date_to
      in: query
      schema:
        type: string
        format: date
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    APIResponseWithMeta:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
    DocumentUpload:
      type: object
      required: [title, files]
      properties:
        title:
          type: string
        doc_type:
          type: string
          enum: [sop, manual, certificate, other]
        description:
          type: string
        files:
          type: array
          items:
            type: string
            format: binary
...
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        doc_type:
          type: string
          enum: [sop, manual, certificate, other]
        description:
          type: string
...
    MaintenanceDocumentUpload:
      type: object
      required: [file]
      properties:
        doc_type:
          type: string
          enum: [certificate, report, other]
        file:
          type: string
          format: binary
...